Got web front end auth to work on Mothra by removing session cookie. Instead, we just gen a UUID on service startup, and hide the priv interactions behind that. Login now redirects to that UUID.
Also reworked some web template code to work better in mothra and abaco. Probably not going to spend a lot of time on abaco compatibility otherwise as it doesn’t seem to support drop down menus.

Fixed issue with IP tables not allowing VM to access the webservice on 192.168.9.1 (thank AI chatbot…).

Fixed several issues with Tor not working, such as such as with the optional tor overlay not backing up the torrc properly, as well as messing up the torrc if run multiple times which .

Fixed issue with restoring the clear net not restarting the “find internet” service.

+++ optional/tor.sh
@@ -14,14 +14,24 @@ outbound()
           sudo cp /etc/iptables/rules.v4 /etc/iptables/rules.v4_back
    fi
 
-   # "Anonymizing Middlebox" from https://gitlab.torproject.org/legacy/trac/-/wikis/doc/TransparentProxy
-   sudo sh -c "( echo \"
+   if [ ! -f "/etc/tor/torrc_prebelagos" ]; then
+      sudo cp /etc/tor/torrc /etc/tor/torrc_prebelagos
+   fi
+
+   # Be careful around appending more than once, we may break tor
+   if [ `grep "TransPort 192.168.9.1:9040" /etc/tor/torrc | wc -l` -eq 0 ]; then
+      # "Anonymizing Middlebox" from https://gitlab.torproject.org/legacy/trac/-/wikis/doc/TransparentProxy
+      sudo sh -c "( echo \"
 VirtualAddrNetworkIPv4 10.192.0.0/10
 AutomapHostsOnResolve 1
 TransPort 192.168.9.1:9040
 DNSPort 192.168.9.1:5353\" >> /etc/tor/torrc )"
+   fi
 
    sudo systemctl restart tor
+
+   # Bypass rules(?) for internal network, so we can like hit our Webservice?
+   sudo iptables -t nat -I PREROUTING -i tap0 -p tcp -d 192.168.0.0/16 -j RETURN
 
    # IP Tables redirect dns to 5353 and everything else to tor
    sudo iptables -t nat -A PREROUTING -i tap0 -p udp --dport 53 -j REDIRECT --to-ports 5353
@@ -36,24 +46,29 @@ inbound()
       install
    fi
 
-   sudo cp /etc/tor/torrc /etc/tor/torrc_prebelagos
+   if [ ! -f "/etc/tor/torrc_prebelagos" ]; then
+      sudo cp /etc/tor/torrc /etc/tor/torrc_prebelagos
+   fi
 
-   if [ $type = 'grid' ]; then
-      sudo sh -c "( echo \"
+   # Be careful around appending more than once, we may break tor
+   if [ `grep "HiddenServicePort 564 192.168.9.3:564" /etc/tor/torrc | wc -l` -eq 0 ]; then
+      if [ $type = 'grid' ]; then
+         sudo sh -c "( echo \"
 HiddenServiceDir /var/lib/tor/hidden_service/
 HiddenServicePort 564 192.168.9.3:564
 HiddenServicePort 567 192.168.9.4:567
 HiddenServicePort 5356 192.168.9.4:5356
 HiddenServicePort 17019 192.168.9.5:17019
 HiddenServicePort 17020 192.168.9.5:17020\" >> /etc/tor/torrc )"
-   else
-      sudo sh -c "( echo \"
+      else
+         sudo sh -c "( echo \"
 HiddenServiceDir /var/lib/tor/hidden_service/
 HiddenServicePort 564 192.168.9.3:564
 HiddenServicePort 567 192.168.9.3:567
 HiddenServicePort 5356 192.168.9.3:5356
 HiddenServicePort 17019 192.168.9.3:17019
 HiddenServicePort 17020 192.168.9.3:17020\" >> /etc/tor/torrc )"
+      fi
    fi
 
    sudo systemctl restart tor
